update and remove files
